Fix Delete functionality and show notifications
